requirements_dev.txt
pip
bump2version
wheel
watchdog
flake8
coverage
Sphinx
sphinx-rtd-theme
twine
pre-commit
pytest
pytest-runner
pytest-cov
setuptools
nbsphinx
ipykernel
jupyter
loguru
scipy
h5netcdf
h5py<3
git+https://github.com/keewis/sphinx-autosummary-accessors
numpydoc
xarray
mypy
bottleneck
xdoctest
matplotlib
